moved ftm processor code to its own file
